requirements.txt
numpy
pillow
tqdm
tensorflow
scikit-learn
requests